Split LLM HTTP provider adapters
- Kept LLMClient as the stable LLMClientProtocol facade while routing HTTP providers to provider-specific adapters.
- Extracted OpenAI-compatible, Anthropic, and Ollama request building, response parsing, streaming EOF policy, and model listing into dedicated adapter files.
- Added a shared HTTP transport/error/model-catalog helper for behavior the old implementation already shared.
- Added adapter-level golden request tests plus streaming EOF and mid-response termination coverage.

R2-4 of PR #725 calls for a character-identical architecture split at the LLM HTTP seam. The goal is to make provider wire rules reviewable independently without changing URLs, headers, request bodies, timeouts, model filtering, streaming sentinel behavior, cancellation surfaces, or API-key scrubbing.

- swift test --filter LLMHTTPAdapterTests
- swift test --filter LLMClientTests
- swift test --filter RoutingLLMClientTests
- swift test --filter LLMServiceTests
